+++ src/option.c
@@ -180,4 +180,5 @@ void set_default_options(void)
    opt_set_int(OPT_PRESERVE_CASE, 0);
    opt_set_str(OPT_GVN_VERBOSE, getenv("NVC_GVN_VERBOSE"));
    opt_set_str(OPT_DCE_VERBOSE, getenv("NVC_GVN_VERBOSE"));
+   opt_set_int(OPT_RANDOM_SEED, get_timestamp_us());
 }